requirements.txt
appnope==0.1.0
asgiref==3.2.7
backcall==0.1.0
decorator==4.4.2
Django==3.0.6
django-autoslug-iplweb==1.9.4.dev0
django-cities==0.6
django-crispy-forms==1.9.0
django-model-utils==4.0.0
django-pdb==0.6.2
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
mysqlclient==1.4.6
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
progressbar2==3.51.3
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-utils==2.4.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
tqdm==4.46.0
traitlets==4.3.3
Unidecode==1.1.1
wcwidth==0.1.9
